02_HR_agent/requirements.txt
fastapi
python-multipart
uvicorn
openai>=1.26.0,<2.0.0 
pydantic
python-dotenv

langchain==0.2.2
langchain-openai==0.1.8
